Consistent hashing function
At the heart of the concept here is that the same container
specification will result in the same container name.  We need
consistent hashing (a la git, docker, etc.) to accomplish this.

The Tainers::hash method provides this.  The test is by no means
exhaustive, but illustrates simple and complex cases.