Tainers.specify method of name management
A relatively thin wrapper around a Tainers::Specification, except
that:
* Tainers helpfully hashes all the params into a hexdigest
* Tainers takes whatever prefix and suffix you provide and, using
  the hash, assembles a deterministic name for the specified
  configuration.

So the Tainers::Specification you get has the name auto-generated
for you.  The prefix and suffix basically let you organize things
in namespaces; prefix defaults to "Tainers", which is guaranteed to
not overlap with your prefixes, as they get forced to lower case.