Tainers::Specification#create method
The #create method is not that different from the #ensure method,
except it will return false in the event that no creation happened.

This means the container already existing, or a "409 Conflict" from
the API, will lead to a false result.

You use #create in cases where you want to react differently to a
successful creation than you would to non-creation, while you
use #ensure in cases where you only care about existence.