Move Specification to Specification::Bare
This prepares for the introduction of a wrapper class on top
of a baseline container manager; the wrapper class will perform
automatic image pulling as necessary, and will wrap the foundational
container management component.

This lets us keep concerns reasonably separated and things reasonably
extensible.

The existing Specification functionality becomes the Bare functionality,
as it does container management and nothing more.